fix: Make error more versatile by changing inner type to String
