+++ src/store/sqlite_store/mod.rs
@@ -1,5 +1,4 @@
 use alloc::collections::BTreeMap;
-use miden_tx::{DataStoreError, TransactionInputs};
 use core::cell::{RefCell, RefMut};
 
 use miden_objects::{
@@ -92,7 +91,6 @@ mod transactions;
 ///     ```
 /// - Thus, if needed you can create a struct representing the json values and use serde_json to
 /// simplify all of the serialization/deserialization logic
-#[derive(Clone)]
 pub struct SqliteStore {
     pub(crate) db: RefCell<Connection>,
 }
@@ -110,82 +108,10 @@ impl SqliteStore {
     }
 
     /// Returns a mutable reference to the internal [Connection] to the SQL DB
-    pub fn store(&self) -> RefMut<'_, Connection> {
+    pub fn db(&self) -> RefMut<'_, Connection> {
         self.db.borrow_mut()
     }
 }
-
-impl<S: Store> miden_tx::DataStore for S {
-    fn get_transaction_inputs(
-        &self,
-        account_id: AccountId,
-        block_num: u32,
-        notes: &[NoteId],
-    ) -> Result<TransactionInputs, DataStoreError> {
-        // First validate that no note has already been consumed
-        let unspent_notes = self
-            .get_input_notes(NoteFilter::Committed)?
-            .iter()
-            .map(|note_record| note_record.id())
-            .collect::<Vec<_>>();
-
-        for note_id in notes {
-            if !unspent_notes.contains(note_id) {
-                return Err(DataStoreError::NoteAlreadyConsumed(*note_id));
-            }
-        }
-
-        // Construct Account
-        let (account, seed) = self.get_account(account_id)?;
-
-        // Get header data
-        let (block_header, _had_notes) = self.get_block_header_by_num(block_num)?;
-
-        let mut list_of_notes = vec![];
-
-        let mut notes_blocks: Vec<u32> = vec![];
-        for note_id in notes {
-            let input_note_record = self.get_input_note(*note_id)?;
-
-            let input_note: InputNote = input_note_record
-                .try_into()
-                .map_err(|err: ClientError| DataStoreError::InternalError(err.to_string()))?;
-
-            list_of_notes.push(input_note.clone());
-
-            let note_block_num = input_note.proof().origin().block_num;
-
-            if note_block_num != block_num {
-                notes_blocks.push(note_block_num);
-            }
-        }
-
-        let notes_blocks: Vec<BlockHeader> = self
-            .get_block_headers(&notes_blocks)?
-            .iter()
-            .map(|(header, _has_notes)| *header)
-            .collect();
-
-        let partial_mmr = build_partial_mmr_with_paths(self, block_num, &notes_blocks)?;
-        let chain_mmr = ChainMmr::new(partial_mmr, notes_blocks)
-            .map_err(|err| DataStoreError::InternalError(err.to_string()))?;
-
-        let input_notes =
-            InputNotes::new(list_of_notes).map_err(DataStoreError::InvalidTransactionInput)?;
-
-        TransactionInputs::new(account, seed, block_header, chain_mmr, input_notes)
-            .map_err(DataStoreError::InvalidTransactionInput)
-    }
-
-    fn get_account_code(&self, account_id: AccountId) -> Result<ModuleAst, DataStoreError> {
-        let (account, _seed) = self.get_account(account_id)?;
-        let module_ast = account.code().module().clone();
-
-        Ok(module_ast)
-    }
-}
-
-
 
 // SQLite implementation of the Store trait
 //
